Stable Overhaul (0.2)
Data can now be saved and loaded. However, new setttingsDefs (previously categoryDefs), does not get detected when loading. Thus an out of range exception occurs. It needs to detect if a new settingsDef is created and initialize it.

On another note, ToolBox hinders greatly in startup. Loading settingsDef data hits hard the more they are. Probably lessen the foreach. Find a way to improve performance!