Stable Overhaul (0.8)
An overhaul of the overhaul. It uses an object as its property data storage. For the most part, It can save and load data and requires only a couple of adapatability relating to removal of thingDef. A little more tweaks and checking of bugs; I can then use it as standard.